get_dimensions placement (#5)
