+++ kitsune/train.py
@@ -7,8 +7,8 @@ from typing import Optional
 import numpy as np
 import torch
 
-from kitsune.data import FileFormat, build_input_data_pipe
-from kitsune.engine import build_feature_mapper, get_dimensions, train_single_epoch
+from kitsune.data import FileFormat, build_input_data_pipe, get_dimensions
+from kitsune.engine import build_feature_mapper, train_single_epoch
 from kitsune.models import FeatureMapper, Kitsune
 from kitsune.scalers import BatchTorchMinMaxScaler
